testContinuity() only in TestMonteCarloBarostat
