+++ tests/TestMonteCarloAnisotropicBarostat.h
@@ -38,10 +38,8 @@
 #include "openmm/System.h"
 #include "openmm/LangevinIntegrator.h"
 #include "openmm/VerletIntegrator.h"
-#include "openmm/CustomVolumeForce.h"
 #include "sfmt/SFMT.h"
 #include "SimTKOpenMMRealType.h"
-#include "TestMonteCarloBarostatContinuity.h"
 #include <iostream>
 #include <vector>
 
@@ -184,86 +182,6 @@ void testIdealGasAxis(int axis) {
         if (!scaleZ) {
             ASSERT(boxZ == 2*initialLength);
         }
-    }
-}
-
-void testContinuity() {
-    // Tests that particles do not get swapped or jump across periodic box
-    // boundaries due to improper handling of atom reordering in the barostat.
-
-    const int numParticles = 64;
-    const double pressure = 1.5;
-    const double pressureInMD = pressure*(AVOGADRO*1e-25); // pressure in kJ/mol/nm^3
-    const double temp = 300.0;
-    const double initialVolume = numParticles*BOLTZ*temp/pressureInMD;
-    const double initialLength = std::pow(initialVolume, 1.0/3.0);
-
-    // Create a gas of particles.
-
-    System system;
-    system.setDefaultPeriodicBoxVectors(Vec3(initialLength, 0, 0), Vec3(0, 0.8*initialLength, 0), Vec3(0, 0, 1.25*initialLength));
-    vector<Vec3> positions(numParticles);
-    OpenMM_SFMT::SFMT sfmt;
-    init_gen_rand(0, sfmt);
-    for (int i = 0; i < numParticles; ++i) {
-        system.addParticle(1.0);
-        positions[i] = Vec3(initialLength*genrand_real2(sfmt), 0.8*initialLength*genrand_real2(sfmt), 1.25*initialLength*genrand_real2(sfmt));
-    }
-    MonteCarloAnisotropicBarostat* barostat = new MonteCarloAnisotropicBarostat(Vec3(pressure, pressure, pressure), temp);
-    system.addForce(barostat);
-
-    // Prevent the box from getting too elongated.
-
-    CustomVolumeForce* volForce = new CustomVolumeForce("(ax-l)^2+(by-l)^2+(cz-l)^2; l=(ax+by+cz)/3");
-    system.addForce(volForce);
-
-    // Add a nonbonded force (to ensure that atom reordering takes place).
-
-    NonbondedForce* pair = new NonbondedForce();
-    pair->setCutoffDistance(2.5);
-    pair->setNonbondedMethod(NonbondedForce::CutoffPeriodic);
-    for (int i = 0; i < numParticles; i++) {
-        pair->addParticle(0.0, 1.0, 0.1);
-    }
-    system.addForce(pair);
-
-    barostat->setDefaultTemperature(temp);
-    LangevinIntegrator integrator(temp, 0.1, 0.005);
-    Context context(system, integrator, platform);
-    context.setPositions(positions);
-
-    vector<Vec3> last, current;
-
-    // Warm up.
-
-    barostat->setFrequency(1);
-    integrator.step(2500);
-
-    // Check with rescaling on.
-
-    for (int i = 0; i < 2500; i++) {
-        integrator.step(1);
-        State state = context.getState(State::Positions);
-        checkContinuity(state, last, current, 0.01);
-    }
-
-    // Check with rescaling on and pressure computation.
-
-    for (int i = 0; i < 2500; i++) {
-        integrator.step(1);
-        barostat->computeCurrentPressure(context);
-        State state = context.getState(State::Positions);
-        checkContinuity(state, last, current, 0.01);
-    }
-
-    // Check with rescaling off and pressure computation.
-
-    barostat->setFrequency(0);
-    for (int i = 0; i < 2500; i++) {
-        integrator.step(1);
-        barostat->computeCurrentPressure(context);
-        State state = context.getState(State::Positions);
-        checkContinuity(state, last, current, 0.01);
     }
 }
 
@@ -715,7 +633,6 @@ int main(int argc, char* argv[]) {
         testIdealGasAxis(0);
         testIdealGasAxis(1);
         testIdealGasAxis(2);
-        testContinuity();
         testMoleculeScaling(true);
         testMoleculeScaling(false);
         testMolecularGas(true);
